fix: resolve CLI not exiting
