+++ src/cli/index.ts
@@ -11,20 +11,20 @@ import { injectGlobals } from "../core/injectGlobals";
 import { logger } from "../core/Logger";
 import { commandFactory } from "./common";
 
-initEventHandlers();
-initEnvVariables();
-
-// global variables can not be injected when running unity tests
-// to not conflict with jest
-
-if (!runtime.isUnityTest) {
-  //logger.mock();
-  injectGlobals().catch((e) => logger.error("could not load corde's globals: ", e));
-}
-
 export const program = new Command();
 
 try {
+  initEventHandlers();
+  initEnvVariables();
+
+  // global variables can not be injected when running unity tests
+  // to not conflict with jest
+
+  if (!runtime.isUnityTest) {
+    logger.mock();
+    injectGlobals();
+  }
+
   // Add basic information with default run all command
   program
     .name("corde")
@@ -34,7 +34,7 @@ try {
 
   commandFactory.loadCommands(program);
 } catch (error) {
-  console.log(error);
+  logger.log(error);
 }
 
 if (process.env.ENV !== "UNITY_TEST" && process.env.ENV !== "E2E_TEST") {
