+++ src/core/injectGlobals.ts
@@ -1,4 +1,10 @@
 import { cordeInternal } from "../api";
+import runtime from "./runtime";
+import * as hooks from "../hooks";
+import * as closures from "../hooks";
+import { expect } from "../expect";
+import { command } from "../command";
+import { BotAPI } from "../api/botAPI";
 
 function getGlobal() {
   return global as any;
@@ -8,12 +14,7 @@ function addToGlobalScope(name: string, value: any) {
   getGlobal()[name] = value;
 }
 
-export async function injectGlobals() {
-  const hooks = await import("../hooks");
-  const closures = await import("../closures");
-  const { expect } = await import("../expect");
-  const { command } = await import("../command");
-
+export function injectGlobals() {
   Object.getOwnPropertyNames(hooks).forEach((hookName) => {
     addToGlobalScope(hookName, (hooks as any)[hookName]);
   });
@@ -26,4 +27,5 @@ export async function injectGlobals() {
   addToGlobalScope("command", command);
   addToGlobalScope("con", command);
   addToGlobalScope("corde", cordeInternal);
+  addToGlobalScope("bot", new BotAPI(runtime.bot));
 }
